Extract some human readable size operations to its own utils file
